Move files to appropriate location
Mistake in last commit
