+++ lib/ecco/event_listener.rb
